Change penalty parameters from Float to Double type
Change `repetitionPenalty` and `lengthPenalty` parameter types from Float to Double for consistency with API expectations and update detection score to use double precision.

+++ modules/watsonx-ai/src/main/java/com/ibm/watsonx/ai/detection/BaseDetectionResponse.java
@@ -11,7 +11,7 @@ abstract class BaseDetectionResponse {
     protected String text;
     protected String detectionType;
     protected String detection;
-    protected float score;
+    protected double score;
 
     public String getText() {
         return text;
@@ -37,11 +37,11 @@ abstract class BaseDetectionResponse {
         this.detection = detection;
     }
 
-    public float getScore() {
+    public double getScore() {
         return score;
     }
 
-    public void setScore(float score) {
+    public void setScore(Double score) {
         this.score = score;
     }
 }
